Move internal.Grpc to internal.AbstractGrpcProtocol
